Integrate Coverage into FeatureDataset (3); Tinker with FeatureDatasetFactoryManager and its factories; Pull dataset location / protocol detection out of NetcdfDataset; Evolve 2D time axis (1); Fix some docs.
