+++ cdm-test/src/test/java/ucar/nc2/util/net/TestMisc.java
@@ -41,7 +41,6 @@ import ucar.httpservices.HTTPFactory;
 import ucar.httpservices.HTTPMethod;
 import ucar.httpservices.HTTPSession;
 import ucar.nc2.util.EscapeStrings;
-import ucar.nc2.util.Misc;
 import ucar.nc2.util.UnitTestCommon;
 import ucar.unidata.test.util.NeedsExternalResource;
 
@@ -120,45 +119,5 @@ public class TestMisc extends UnitTestCommon
 
     }
 
-    protected boolean protocheck(String path, String expected)
-    {
-        if(expected == null)
-            expected = "";
-        List<String> protocols = Misc.getProtocols(path);
-        StringBuilder buf = new StringBuilder();
-        for(String s : protocols) {
-            buf.append(s);
-            buf.append(":");
-        }
-        String result = buf.toString();
-        boolean ok = expected.equals(result);
-        System.err.printf("path=|%s| result=|%s| pass=%s\n",
-            path, result, (ok ? "true" : "false"));
-        System.err.flush();
-        return ok;
-    }
-
-    @Test
-    public void
-    testGetProtocols()
-    {
-        String tag = "TestMisc.testGetProtocols";
-        Assert.assertTrue(tag, protocheck("http://server/thredds/dodsC/", "http:"));
-        Assert.assertTrue(tag, protocheck("dods://thredds-test.unidata.ucar.edu/thredds/dodsC/grib/NCEP/NAM/CONUS_12km/best", "dods:"));
-        Assert.assertTrue(tag, protocheck("dap4://ucar.edu:8080/x/y/z", "dap4:"));
-        Assert.assertTrue(tag, protocheck("dap4:https://ucar.edu:8080/x/y/z", "dap4:https:"));
-        Assert.assertTrue(tag, protocheck("file:///x/y/z", "file:"));
-        Assert.assertTrue(tag, protocheck("file://c:/x/y/z", "file:"));
-        Assert.assertTrue(tag, protocheck("file:c:/x/y/z", "file:"));
-        Assert.assertTrue(tag, protocheck("file:/blah/blah/some_file_2014-04-13_16:00:00.nc.dds", "file:"));
-        Assert.assertTrue(tag, protocheck("/blah/blah/some_file_2014-04-13_16:00:00.nc.dds", ""));
-        Assert.assertTrue(tag, protocheck("c:/x/y/z", null));
-        Assert.assertTrue(tag, protocheck("x::a/y/z", null));
-        Assert.assertTrue(tag, protocheck("x::/y/z", null));
-        Assert.assertTrue(tag, protocheck("::/y/z", ""));
-        Assert.assertTrue(tag, protocheck("dap4:&/y/z", null));
-        Assert.assertTrue(tag, protocheck("file:x/z::a", "file:"));
-        Assert.assertTrue(tag, protocheck("x/z::a", null));
-    }
 
 }
