+++ cdm/src/main/java/ucar/nc2/dataset/NetcdfDataset.java
@@ -40,17 +40,12 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.EnumSet;
 import java.util.Formatter;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
 
-import org.apache.http.Header;
 import org.jdom2.Element;
 import thredds.client.catalog.ServiceType;
 import thredds.client.catalog.tools.DataFactory;
-import ucar.httpservices.HTTPFactory;
-import ucar.httpservices.HTTPMethod;
 import ucar.ma2.Array;
 import ucar.ma2.DataType;
 import ucar.ma2.InvalidRangeException;
@@ -72,26 +67,19 @@ import ucar.nc2.ncml.NcMLWriter;
 import ucar.nc2.stream.CdmRemote;
 import ucar.nc2.util.CancelTask;
 import ucar.nc2.util.CancelTaskImpl;
-import ucar.nc2.util.EscapeStrings;
-import ucar.nc2.util.Misc;
 import ucar.nc2.util.cache.FileCache;
 import ucar.nc2.util.cache.FileFactory;
 import ucar.unidata.util.StringUtil2;
-import ucar.unidata.util.Urlencoded;
 
 /**
  * NetcdfDataset extends the netCDF API, adding standard attribute parsing such as
  * scale and offset, and explicit support for Coordinate Systems.
  * A NetcdfDataset wraps a NetcdfFile, or is defined by an NcML document.
  * <p/>
- * <p> Be sure to close the dataset when done, best practice is to enclose in a try/finally block:
+ * <p> Be sure to close the dataset when done, best practice is to use try-with-resource:
  * <pre>
- * NetcdfDataset ncd = null;
- * try {
- *   ncd = NetcdfDataset.openDataset(fileName);
+ * try (NetcdfDataset ncd = NetcdfDataset.openDataset(fileName)) {
  *   ...
- * } finally {
- *   ncd.close();
  * }
  * </pre>
  * <p/>
@@ -673,56 +661,9 @@ public class NetcdfDataset extends ucar.nc2.NetcdfFile {
     if (orgLocation == null)
       throw new IOException("NetcdfDataset.openFile: location is null");
 
-    // Canonicalize the location
     String location = StringUtil2.replace(orgLocation.trim(), '\\', "/");
-    List<String> allprotocols = Misc.getProtocols(location);
-
-    String trueurl = location;
-    String leadprotocol;
-    if (allprotocols.size() == 0) {
-      leadprotocol = "file";  // The location has no leading protocols, assume file:
-    } else {
-      leadprotocol = allprotocols.get(0);
-    }
-
-    // Priority in deciding
-    // the service type is as follows.
-    // 1. "protocol" tag in fragment
-    // 2. leading protocol
-    // 3. path extension
-    // 4. contact the server (if defined)
-
-    // remove any trailing query or fragment
-    String fragment = null;
-    int pos = trueurl.lastIndexOf('#');
-    if (pos >= 0) {
-      fragment = trueurl.substring(pos + 1, trueurl.length());
-      trueurl = trueurl.substring(0, pos);
-    }
-    pos = location.lastIndexOf('?');
-    if (pos >= 0) {
-      trueurl = trueurl.substring(0, pos);
-    }
-
-    ServiceType svctype = null;
-    if (fragment != null)
-      svctype = searchFragment(fragment);
-
-    if (svctype == null) // See if leading protocol tells us how to interpret
-      svctype = decodeLeadProtocol(leadprotocol);
-
-    if (svctype == null) {
-      //There are several possibilities at this point; all of which
-      // require further info to disambiguate
-      //  - we have file://<path> or file:<path>; we need to see if
-      //    the extension can help, otherwise, start defaulting.
-      //  - we have a simple url: e.g. http://... ; contact the server
-      if (leadprotocol.equals("file")) {
-        svctype = decodePathExtension(trueurl); // look at the path extension
-      } else {
-        svctype = disambiguateHttp(trueurl);
-      }
-    }
+    DatasetUrl durl = DatasetUrl.findDatasetUrl(location);
+    ServiceType svctype = durl.serviceType;
 
     if (svctype == ServiceType.OPENDAP)
       return acquireDODS(cache, factory, hashKey, location, buffer_size, cancelTask, spiObject);
@@ -734,10 +675,7 @@ public class NetcdfDataset extends ucar.nc2.NetcdfFile {
       return acquireDap4(cache, factory, hashKey, location, buffer_size, cancelTask, spiObject);
 
     else if (svctype == ServiceType.NCML) {
-      // If lead protocol was null and then pretend it was a file
-      // Note that technically, this should be 'file://'
-      String url = (allprotocols.size() == 0 ? "file:" + trueurl : location);
-      return acquireNcml(cache, factory, hashKey, url, buffer_size, cancelTask, spiObject);
+      return acquireNcml(cache, factory, hashKey, durl.trueurl, buffer_size, cancelTask, spiObject);
 
     } else if (svctype == ServiceType.THREDDS) {
       Formatter log = new Formatter();
@@ -754,240 +692,16 @@ public class NetcdfDataset extends ucar.nc2.NetcdfFile {
       throw new IOException("Unknown service type: " + svctype.toString());
     }
 
-    // Next to last resort: look in the cache
+    // Next to last resort: find in the cache
     if (cache != null) {
       if (factory == null)
         factory = defaultNetcdfFileFactory;
       return (NetcdfFile) cache.acquire(factory, hashKey, location, buffer_size, cancelTask, spiObject);
     }
 
-    // Last resort: try to open as a file
+    // Last resort: try to open as a file or remote file
     return NetcdfFile.open(location, buffer_size, cancelTask, spiObject);
   }
-
-
-  /**
-   * Check path extension; assumes no query or fragment
-   *
-   * @param path the path to examine for extension
-   * @return ServiceType inferred from the extension or null
-   */
-  static ServiceType decodePathExtension(String path) {
-    // Look at the path extensions
-    if (path.endsWith(".dds") || path.endsWith(".das") || path.endsWith(".dods"))
-      return ServiceType.OPENDAP;
-
-    if (path.endsWith(".dmr") || path.endsWith(".dap") || path.endsWith(".dsr"))
-      return ServiceType.DAP4;
-
-    if (path.endsWith(".xml") || path.endsWith(".ncml"))
-      return ServiceType.NCML;
-    return null;
-  }
-
-  /*
-   * Attempt to map a lead url protocol url to a service type
-   * (see thredds.catalog.ServiceType).
-   * Possible service types should include at least the following.
-   * <ol>
-   * <li> OPENDAP (DAP2 protocol)
-   * <li> DAP4 (DAP4 protocol)
-   * <li> CdmRemote (remote ncstream)
-   * </ol>
-   *
-   * @param protocol The leading protocol
-   * @return ServiceType indicating how to handle the url, or null.
-   */
-  @Urlencoded
-  static ServiceType decodeLeadProtocol(String protocol) throws IOException {
-    if (protocol.equals("dods"))
-      return ServiceType.OPENDAP;
-
-    else if (protocol.equals("dap4"))
-      return ServiceType.DAP4;
-
-    else if (protocol.equals("httpserver") || protocol.equals("nodods"))
-       return ServiceType.HTTPServer;
-
-     else if (protocol.equals(CdmRemote.PROTOCOL))
-      return ServiceType.CdmRemote;
-
-    else if (protocol.equals(DataFactory.PROTOCOL)) //thredds
-      return ServiceType.THREDDS;
-
-    return null;
-  }
-
-  /**
-   * If the URL alone is not sufficient to disambiguate the location,
-   * then this method will attempt to do a specific kind of request on
-   * the server, typically a HEAD call using the URL.
-   * It looks for the header "Content-Description"
-   * and uses it value (e.g. "ncstream" or "dods", etc)
-   * in order to disambiguate.
-   *
-   * @param location the url to disambiguate
-   * @return ServiceType indicating how to handle the url
-   */
-  @Urlencoded
-  static private ServiceType disambiguateHttp(String location) throws IOException {
-    // aggregation cache files are of form
-    // http://www.esrl.noaa.gov/psd/thredds/dodsC/Datasets/ncep.reanalysis2.dailyavgs/pressure/air.1981.nc#320092027
-
-    ServiceType result = checkIfDods(location); // dods
-    if (result != null)
-      return result;
-
-    result = checkIfDap4(location); // dap4
-    if (result != null)
-      return result;
-
-    try (HTTPMethod method = HTTPFactory.Head(location)) {
-      int statusCode = method.execute();
-      if (statusCode >= 300) {
-        if (statusCode == 401)
-          throw new IOException("Unauthorized to open dataset " + location);
-        else
-          throw new IOException(location + " is not a valid URL, return status=" + statusCode);
-      }
-
-      Header h = method.getResponseHeader("Content-Description");
-      if ((h != null) && (h.getValue() != null)) {
-        String v = h.getValue();
-        if (v.equalsIgnoreCase("ncstream"))
-          return ServiceType.CdmRemote;
-      }
-
-      return null;
-    }
-  }
-
-  // not sure what other opendap servers do, so fall back on check for dds
-  static private ServiceType checkIfDods(String location) throws IOException {
-    int len = location.length();
-    // Strip off any trailing .dds, .das, or .dods
-    if (location.endsWith(".dds"))
-      location = location.substring(0, len - ".dds".length());
-    if (location.endsWith(".das"))
-      location = location.substring(0, len - ".das".length());
-    if (location.endsWith(".dods"))
-      location = location.substring(0, len - ".dods".length());
-    // Opendap assumes that the caller has properly escaped the url
-    try (
-      // For some reason, the head method is not using credentials
-      // method = session.newMethodHead(location + ".dds");
-      HTTPMethod method = HTTPFactory.Get(location + ".dds")) {
-
-      int status = method.execute();
-      if (status == 200) {
-        Header h = method.getResponseHeader("Content-Description");
-        if ((h != null) && (h.getValue() != null)) {
-          String v = h.getValue();
-          if (v.equalsIgnoreCase("dods-dds") || v.equalsIgnoreCase("dods_dds"))
-            return ServiceType.OPENDAP;
-          else
-            throw new IOException("OPeNDAP Server Error= " + method.getResponseAsString());
-        }
-      }
-      if (status == 401)
-        throw new IOException("Unauthorized to open dataset " + location);
-
-      // not dods
-      return null;
-    }
-  }
-
-  // check for dmr
-  static private ServiceType checkIfDap4(String location) throws IOException {
-    // Strip off any trailing DAP4 prefix
-    if (location.endsWith(".dap"))
-      location = location.substring(0, location.length() - ".dap".length());
-    else if (location.endsWith(".dmr"))
-      location = location.substring(0, location.length() - ".dmr".length());
-    else if (location.endsWith(".dsr"))
-      location = location.substring(0, location.length() - ".dsr".length());
-    try (HTTPMethod method = HTTPFactory.Get(location + ".dmr")) {
-      int status = method.execute();
-      if (status == 200) {
-        Header h = method.getResponseHeader("Content-Type");
-        if ((h != null) && (h.getValue() != null)) {
-          String v = h.getValue();
-          if (v.startsWith("application/vnd.opendap.org"))
-            return ServiceType.DAP4;
-        }
-      }
-      if (status == 401)
-        throw new IOException("Unauthorized to open dataset " + location);
-
-      // not dods
-      return null;
-    }
-  }
-
-  /**
-   * Given a location look for
-   * markers indicated which protocol to use
-   *
-   * @param fragment the fragment is to be examined
-   * @return The discovered ServiceType, or null
-   */
-  static ServiceType searchFragment(String fragment) {
-    if (fragment.length() == 0)
-      return null;
-    Map<String, String> map = parseFragment(fragment);
-    if (map == null) return null;
-    String protocol = map.get("protocol");
-    if (protocol != null) {
-      if (protocol.equalsIgnoreCase("dap")
-              || protocol.equalsIgnoreCase("dods"))
-        return ServiceType.OPENDAP;
-      if (protocol.equalsIgnoreCase("dap4"))
-        return ServiceType.DAP4;
-      if (protocol.equalsIgnoreCase("cdmremote"))
-        return ServiceType.CdmRemote;
-      if (protocol.equalsIgnoreCase("thredds"))
-        return ServiceType.THREDDS;
-      if (protocol.equalsIgnoreCase("ncmdl"))
-        return ServiceType.NCML;
-    }
-    return null;
-  }
-
-  /**
-   * Given the fragment part of a url, see if it
-   * parses as name=value pairs separated by '&'
-   * (same as query part).
-   *
-   * @param fragment the fragment part of a url
-   * @return a map of the name value pairs (possibly empty),
-   * or null if the fragment does not parse.
-   */
-
-  static Map<String, String> parseFragment(String fragment) {
-    Map<String, String> map = new HashMap<>();
-    if (fragment != null && fragment.length() >= 0) {
-      if (fragment.charAt(0) == '#')
-        fragment = fragment.substring(1);
-      String[] pairs = fragment.split("[ \t]*[&][ \t]*");
-      for (String pair : pairs) {
-        String[] pieces = pair.split("[ \t]*[=][ \t]*");
-        switch (pieces.length) {
-          case 1:
-            map.put(EscapeStrings.unescapeURL(pieces[0]).toLowerCase(),
-                    "true");
-            break;
-          case 2:
-            map.put(EscapeStrings.unescapeURL(pieces[0]).toLowerCase(),
-                    EscapeStrings.unescapeURL(pieces[1]).toLowerCase());
-            break;
-          default:
-            return null; // does not parse
-        }
-      }
-    }
-    return map;
-  }
-
 
   //////////////////////////////////////////////////
 
@@ -1092,6 +806,8 @@ public class NetcdfDataset extends ucar.nc2.NetcdfFile {
     }
 
   }
+
+  ////////////////////////////////////////////////////////////////////////////////////
 
   static private NetcdfFile acquireNcml(FileCache cache, FileFactory factory, Object hashKey,
                                         String location, int buffer_size, ucar.nc2.util.CancelTask cancelTask, Object spiObject) throws IOException {
