+++ docs/website/netcdf-java/NetcdfUsers.adoc
@@ -1,13 +1,9 @@
 Netcdf Users
 ============
 
- +
- image:NetcdfUsers.jpg[image]
+image:NetcdfUsers.jpg[image]
 
- 
-
-The ``brightness'' of this image is proportional to the number of netcdf
-users in that geographic area.
+The *brightness* of this image is proportional to the number of netcdf users in that geographic area.
 
 _Created by the **D**epartment for **A**vailable **M**etrics on
 **N**etcdf, **L**DM, **I**DV, and **E**arlier **S**oftware._
