+++ docs/website/netcdf-java/reference/RuntimeLoading.adoc
@@ -1,6 +1,3 @@
-Runtime Loading
-===============
-
 = Runtime Loading
 
 These are the various classes that can be plugged in at runtime:
@@ -98,9 +95,7 @@ about lookup tables.
 
 The file must be a link:formats/BufrTables.html[BUFR table lookup file].
 
-'''''
-
- +
+''''
 
 == Runtime Configuration
 
@@ -110,9 +105,10 @@ library an XML configuration file. Note that your application must call
 
 The configuration file looks like this:
 
---------------------------------------------------------------------------------------------------------------------------
- <?xml version="1.0"?>
- <nj22Config>
+[source,xml]
+----
+<?xml version="1.0"?>
+<nj22Config>
 
 1) <ioServiceProvider  class="edu.univ.ny.stuff.FooFiles"/>
 2) <coordSysBuilder convention="foo" class="test.Foo"/>
@@ -129,7 +125,7 @@ The configuration file looks like this:
      <libraryName>netcdf4</libraryName>
    </Netcdf4Clibrary>
 </nj22Config>
---------------------------------------------------------------------------------------------------------------------------
+----
 
 1.  Loads an *IOServiceProvider* with the given class name
 2.  Loads a *CoordSysBuilderIF* with the given class name, which looks
@@ -157,7 +153,8 @@ library. From your application, you can pass a *java.io.InputStream* (or
 http://www.jdom.org/[JDOM] element) to
 **ucar.nc2.util.xml.RuntimeConfigParser**, as in the following examples:
 
----------------------------------------------------------------------------
+[source,java]
+----
   // Example 1: read from file
   StringBuffer errlog = new StringBuffer();
   FileInputStream fis = new FileInputStream( filename);   
@@ -181,42 +178,37 @@ http://www.jdom.org/[JDOM] element) to
   Element elem = root.getChild("nj22Config");
   if (elem != null)
     ucar.nc2.util.RuntimeConfigParser.read( elem, errlog);
----------------------------------------------------------------------------
+----
 
 For example, the ToolsUI application allows you to specify this file on
 the command line with the -**nj22Config** parameter:
 
--------------------------------------------------------------------------------------
-   public void main(String[] args) {
+[source,java]
+----
+public void main(String[] args) {
 
-      for (int i = 0; i < args.length; i++) {
-        if (args[i].equalsIgnoreCase("-nj22Config") && (i < args.length-1)) {
-          String runtimeConfig = args[i+1];
-          i++;
-          try {
-            StringBuffer errlog = new StringBuffer();
+  for (int i = 0; i < args.length; i++) {
+    if (args[i].equalsIgnoreCase("-nj22Config") && (i < args.length-1)) {
+      String runtimeConfig = args[i+1];
+      i++;
+      try {
+        StringBuffer errlog = new StringBuffer();
+        FileInputStream fis = new FileInputStream( runtimeConfig);
+        ucar.nc2.util.xml.RuntimeConfigParser.read( fis, errlog);
+        System.out.println( errlog);
 
-    FileInputStream fis = new FileInputStream( runtimeConfig);
- ucar.nc2.util.xml.RuntimeConfigParser.read( fis, errlog);
-            System.out.println( errlog);
-          } catch (IOException ioe) {
-            System.out.println( "Error reading "+runtimeConfig+"="+ioe.getMessage());
-          }
-        }
-      }
-    ...
--------------------------------------------------------------------------------------
+     } catch (IOException ioe) {
+       System.out.println( "Error reading "+runtimeConfig+"="+ioe.getMessage());
+     }
+   }
+ }
+...
+----
 
 If none is specified on the command line, it will look for the XML
 document in *$USER_HOME/.unidata/nj22Config.xml.*
 
-'''''
-
- 
-
 == Runtime Loading of IOSP using javax.imageio.spi.ServiceRegistry
-
-(as of version 4.3.9)
 
 You can create an IOSP and have it discovered at runtime automatically.
 
@@ -237,14 +229,9 @@ containing the line:
 com.mycompany.MyIOSP
 --------------------
 +
-See:
-
-___________________________________________________________________________________
-http://docs.oracle.com/javase/1.4.2/docs/api/javax/imageio/spi/ServiceRegistry.html
+See: http://docs.oracle.com/javase/1.4.2/docs/api/javax/imageio/spi/ServiceRegistry.html
 
 (thanks to Tom Kunicki at USGS for this contribution)
-___________________________________________________________________________________
 
 '''''
-
-image:../nc.gif[image] This document was last updated on March 2015
+image:../nc.gif[image] This document was last updated October 2015
