+++ docs/website/netcdf-java/reference/stream/CdmRemote.adoc
@@ -1,7 +1,5 @@
 = CDM Remote Web Service
-:linkcss:
-:stylesheet: ../../cdm.css
-:tdsDocs: ../../../tds/reference/services/
+:tdsDocs: ../../../tds/reference/services
 
 CDM Remote is a web service implemented in the CDM library (client) and TDS (server), providing remote access to CDM datasets, using
 link:NcStream.html[ncstream] as the on-the-wire protocol. It provides access at the NetcdfFile and FeatureDataset levels of the CDM API, so there are
@@ -9,8 +7,8 @@ two levels of services:
 
 1.  *CdmRemote* provides index subsetting on remote CDM datasets, with the same functionality that *NetcdfFile* provides for
 CDM local datasets and *DODSNetcdfFile* provides for remote OPeNDAP datasets. CdmRemote supports the full CDM data model.
-2.  *link:CdmrFeature.html[CdmrFeature]* provides coordinate subsetting on remote CDM Feature Datasets, with similar functionality to WCS and
-Unidata's link:{tdsDocs}NetcdfSubsetServiceReference.html[_NetCDF Subset Service_] __(NCSS)__. This API and protocol is redone
+2.  *link:CdmrFeature.adoc[CdmrFeature]* provides coordinate subsetting on remote CDM Feature Datasets, with similar functionality to WCS and
+Unidata's link:{tdsDocs}/NetcdfSubsetServiceReference.html[NetCDF Subset Service] (NCSS). This API and protocol is redone
 in version 5.0.
 
 The CDM Remote services and protocol are experimental and should not be used outside of the CDM stack for now.
@@ -80,12 +78,13 @@ messages.
 
 === Java
 
-*ucar.nc2.stream.CdmRemote* is a subclass of *NetcdfFile* which provides index subsetting on remote CDM datasets. *NetcdfDataset.openOrAcquireFile()*
-looks for the _cdmremote:url_ prefix, and instantiates a *CdmRemote* object. The url must be an endpoint for a cdmremote service.
+*ucar.nc2.stream.CdmRemote* is a subclass of *NetcdfFile* which provides index subsetting on remote CDM datasets.
+*NetcdfDataset.openOrAcquireFile()* sees the _cdmremote:url_ prefix, and instantiates a *CdmRemote* object. The url must be an endpoint for a _cdmremote_ service.
+ToolsUI now preferentially uses _cdmremote_ service when it is available.
 
 * upon opening, *req=header* is called and the NetcdfFile objects are read from the response
-* when data is requested, and the data are not already stored (immediate mode). then *req=data* is called.
-* from the client POV, the dataset looks exactly the same as any other NetcdfFile.
+* when data is requested, and the data are not already stored locally (immediate mode), then *req=data* is called.
+* from the client POV, the dataset is exactly the same as any other *NetcdfFile*.
 
 CDM datasets can also be serialized and stored in a disk file, and then read as a "normal" netCDF file by the CDM library. In this sense *ncstream*
 becomes another format for NetCDF files, alongside netCDF-3 and netCDF-4.
@@ -100,6 +99,6 @@ Python
 
 '''''
 
-image:../../nc.gif[image] This document was last updated August 2015
+image:../../nc.gif[image] This document was last updated October 2015
